parent shape class
added a parent shape class. Now I just need to make sure the comments are all in their proper place for later use.
x is now used to flip on the x axis
y is now used to flip on the y axis
f is used to flip on both axes.
scrollwheel now changes the amount of iterations each item has. (assuming the item is highlighted somehow.)