misc changes
so on the backend, now it works so that you only need to have the self.getPoints function in order to draw regardless of the draw mode.

Circles can now have their curvature changed with a LCTRL + scrollwheel. The math was gotten from chatGPT and I have no idea how it works lmao.

I added a variety of notes on how things work throughout the lines.py file.
I also deleted all of the old code for Circles and Lines.